+++ src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/requests/send/payments/SendInvoice.kt
@@ -5,12 +5,17 @@ import com.github.insanusmokrassar.TelegramBotAPI.bot.RequestsExecutor
 import com.github.insanusmokrassar.TelegramBotAPI.requests.send.abstracts.SendMessageRequest
 import com.github.insanusmokrassar.TelegramBotAPI.types.*
 import com.github.insanusmokrassar.TelegramBotAPI.types.buttons.InlineKeyboardMarkup
-import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.Message
+import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.*
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.TelegramBotAPIMessageDeserializationStrategy
+import com.github.insanusmokrassar.TelegramBotAPI.types.message.content.media.PhotoContent
+import com.github.insanusmokrassar.TelegramBotAPI.types.message.payments.InvoiceOfPayment
 import com.github.insanusmokrassar.TelegramBotAPI.types.payments.LabeledPrice
 import com.github.insanusmokrassar.TelegramBotAPI.types.payments.LabeledPricesSerializer
 import com.github.insanusmokrassar.TelegramBotAPI.types.payments.abstracts.*
 import kotlinx.serialization.*
+
+private val invoiceMessageSerializer: DeserializationStrategy<ContentMessage<InvoiceOfPayment>>
+    = TelegramBotAPIMessageDeserializationStrategyClass()
 
 /**
  * @param providerData - JSON-ENCODED FIELD
@@ -62,10 +67,10 @@ data class SendInvoice(
     DisableNotification,
     ReplyMessageId,
     ReplyMarkup,
-    SendMessageRequest<Message> {
+    SendMessageRequest<ContentMessage<InvoiceOfPayment>> {
     override fun method(): String = "sendInvoice"
-    override val resultDeserializer: DeserializationStrategy<Message>
-        get() = TelegramBotAPIMessageDeserializationStrategy
+    override val resultDeserializer: DeserializationStrategy<ContentMessage<InvoiceOfPayment>>
+        get() = invoiceMessageSerializer
     override val requestSerializer: SerializationStrategy<*>
         get() = serializer()